Move GUI and GUIRenderer to src/
