+++ src/Program.js
@@ -1,3 +1,6 @@
+/**
+ * @todo Add getters
+ */
 export default function Program(program, vertexShader, fragmentShader) {
 	this.program = program;
 	this.vertexShader = vertexShader;
