+++ src/RendererManager.js
@@ -2,6 +2,9 @@ import {NotImplementedError} from "./errors/index.js";
 import Instance from "./Instance.js";
 import WebGLRenderer from "./WebGLRenderer.js";
 
+/**
+ * @todo Convert to function constructor
+ */
 export default class RendererManager {
 	/**
 	 * @param {Instance} instance
