+++ src/cameras/index.js
@@ -1,2 +1,1 @@
-/** @todo Expose `Camera` as well? */
-export {default as OrthographicCamera} from "./OrthographicCamera.js";
+export {OrthographicCamera} from "./OrthographicCamera.js";
