+++ src/gui/Subcomponent.js
@@ -7,4 +7,13 @@ import {Vector2} from "../math/index.js";
  *    uv: Vector2
  * }}
  */
-export function Subcomponent({offset, size, uv}) {}
+export function Subcomponent({offset, size, uv}) {
+	/** @returns {Vector2} */
+	this.getOffset = () => offset;
+
+	/** @returns {Vector2} */
+	this.getSize = () => size;
+
+	/** @returns {Vector2} */
+	this.getUV = () => uv;
+}
