+++ src/gui/index.js
@@ -1,6 +1,8 @@
-export {Layer} from "./Layer.js";
 export {Component} from "./components/Component.js";
 export {DynamicComponent} from "./components/DynamicComponent.js";
 export {StructuralComponent} from "./components/StructuralComponent.js";
 export {VisualComponent} from "./components/VisualComponent.js";
+export {GUI} from "./GUI.js";
+export {GUIRenderer} from "./GUIRenderer.js";
+export {Layer} from "./Layer.js";
 export {Subcomponent} from "./Subcomponent.js";
